Move useApiClient into composables
